Code update v1.0.12

+++ src/hooks/useCalendarStore.js
@@ -1,17 +1,26 @@
 import { useDispatch, useSelector } from 'react-redux';
+import Swal from 'sweetalert2';
 import {
   onAddNewEvent,
   onClearActiveEvent,
   onDeleteEvent,
+  onLoadEvents,
   onSetActiveEvent,
   onUpdateEvent,
 } from '../store';
 import { useUiStore } from './useUiStore';
 
+import { useCallback } from 'react';
+import { calendarApi } from '../api';
+import { parseDates } from '../helpers';
+
 export const useCalendarStore = () => {
   const dispatch = useDispatch();
-  const { events, activeEvent } = useSelector((state) => state.calendar);
-  const { isDateModalOpen } = useUiStore();
+  const { events, activeEvent, isLoadingEvents } = useSelector(
+    (state) => state.calendar
+  );
+  const { user } = useSelector((state) => state.auth);
+  const { isDateModalOpen, closeDateModal } = useUiStore();
 
   const setActiveEvent = (calendarEvent) => {
     dispatch(onSetActiveEvent(calendarEvent));
@@ -22,32 +31,72 @@ export const useCalendarStore = () => {
   };
 
   const startSavingEvent = async (calendarEvent) => {
-    //TODO: Reach backend with info
-    //TODO: All good from backend response
+    try {
+      const eventWithUser = { ...calendarEvent, user: user.uid };
 
-    if (calendarEvent._id) {
-      //updating
-      dispatch(onUpdateEvent({ ...calendarEvent }));
-    } else {
-      //creating
-      dispatch(onAddNewEvent({ ...calendarEvent, _id: new Date().getTime() }));
+      if (calendarEvent.id) {
+        // Updating Event
+        await calendarApi.put(`/events/${calendarEvent.id}`, eventWithUser);
+        dispatch(onUpdateEvent({ ...calendarEvent }));
+        closeDateModal();
+        return;
+      }
+
+      // Creating Event
+      const { data } = await calendarApi.post('/events', eventWithUser);
+      const parsedEvent = parseDates([data.event])[0];
+
+      dispatch(
+        onAddNewEvent({
+          ...parsedEvent,
+          user: {
+            _id: user.uid,
+            name: user.name,
+          },
+        })
+      );
+      closeDateModal();
+    } catch (error) {
+      console.log(error);
+      Swal.fire('Error saving', error.response?.data?.msg, 'error');
     }
   };
 
-  const startDeletingEvent = () => {
-    //TODO REACH BACKEND
-    dispatch(onDeleteEvent());
+  const startDeletingEvent = async () => {
+    if (!activeEvent || !activeEvent.id) return;
+    try {
+      await calendarApi.delete(`/events/${activeEvent.id}`);
+      dispatch(onDeleteEvent());
+    } catch (error) {
+      console.log(error);
+      Swal.fire('Error deleting', error.response?.data?.msg, 'error');
+    }
   };
+
+  const startLoadingEvents = useCallback(async () => {
+    try {
+      const { data } = await calendarApi.get('/events');
+      const events = parseDates(data.events);
+
+      dispatch(onLoadEvents(events));
+    } catch (error) {
+      console.log('Error loading events', error);
+      Swal.fire('Error', 'The events could not be loaded', 'error');
+    }
+  }, [dispatch]);
+
   return {
     //Props
     activeEvent,
     events,
     isDateModalOpen,
+    isLoadingEvents,
 
     //Methods
-    startDeletingEvent,
-    setActiveEvent,
-    startSavingEvent,
     clearActiveEvent,
+    setActiveEvent,
+    startDeletingEvent,
+    startLoadingEvents,
+    startSavingEvent,
   };
 };
